fw/: reduce image width from 100 to 80 square pixels
To achieve square 100 pixels (1:6.25 image aspect ratio), very vigorous
movement is needed. 80 pixels (still 1:5) are easier. Initial estimates
suggested a usable range of 1:4-1:5, so this is still good.

Note that we may want to change the aspect ratio of the pixels themselves
in the future (for smoother letters), so the pixel count may go up again.

80 pixels correspond to 11 characters in a 7x5 font with 2 pixels
separation between characters. We may want to go up to 82 pixels
to allow a 12th character - or use a proportional font.